Read inviting member from env
