+++ env.example
@@ -1,4 +1,4 @@
 SCREENING_AUTHORITY_SEED=//Alice
-ENDOWMENT=200
-PORT=4000
+INVITING_MEMBER_ID=0
+PORT=4004
 PROVIDER=ws://localhost:9944/
